requirements.txt
openai>=2.0.0
google-genai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.1
jsonschema>=4.22.0

